requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Limiter==3.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_resources==6.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.14.1
ordered-set==4.1.0
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
sniffio==1.3.1
SQLAlchemy==2.0.28
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
wrapt==1.16.0
